requirements.txt
# backend/requirements.txt

Django>=4.0,<5.0
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
dj-database-url>=1.0,<2.0
django-filter>=23.0,<24.0
drf-yasg>=1.21,<1.22
gunicorn>=20.0,<21.0
djangorestframework-simplejwt>=5.0,<6.0 
django-cors-headers>=3.0,<4.0
Pillow>=9.0,<11.0
django-extensions==4.1
pygraphviz==1.14
